Move operations-processors to cache dir.
All operation processors are specific to caches.
