+++ lib/orbit-common.js
@@ -5,10 +5,10 @@ import Serializer from 'orbit-common/serializer';
 import Source from 'orbit-common/source';
 import MemorySource from 'orbit-common/memory-source';
 import Transaction from 'orbit-common/transaction';
-import OperationProcessor from 'orbit-common/operation-processors/operation-processor';
-import CacheIntegrityProcessor from 'orbit-common/operation-processors/cache-integrity-processor';
-import DeletionTrackingProcessor from 'orbit-common/operation-processors/deletion-tracking-processor';
-import SchemaConsistencyProcessor from 'orbit-common/operation-processors/schema-consistency-processor';
+import OperationProcessor from 'orbit-common/cache/operation-processors/operation-processor';
+import CacheIntegrityProcessor from 'orbit-common/cache/operation-processors/cache-integrity-processor';
+import DeletionTrackingProcessor from 'orbit-common/cache/operation-processors/deletion-tracking-processor';
+import SchemaConsistencyProcessor from 'orbit-common/cache/operation-processors/schema-consistency-processor';
 import { toIdentifier, parseIdentifier } from 'orbit-common/lib/identifiers';
 import {
   operationType,
