+++ lib/orbit-common/cache.js
@@ -4,8 +4,8 @@ import Evented from 'orbit/evented';
 import { Class, clone, expose, isArray, isObject, isNone } from 'orbit/lib/objects';
 import { eq } from 'orbit/lib/eq';
 import { OperationNotAllowed } from './lib/exceptions';
-import CacheIntegrityProcessor from './operation-processors/cache-integrity-processor';
-import SchemaConsistencyProcessor from './operation-processors/schema-consistency-processor';
+import CacheIntegrityProcessor from './cache/operation-processors/cache-integrity-processor';
+import SchemaConsistencyProcessor from './cache/operation-processors/schema-consistency-processor';
 import Transform from 'orbit/transform';
 import TransformLog from 'orbit/transform-log';
 import { diffs } from 'orbit/lib/diffs';
