+++ lib/orbit-common/cache/operation-processors/cache-integrity-processor.js
@@ -1,8 +1,8 @@
 import { Class, isArray, isObject, isNone } from 'orbit/lib/objects';
 import Operation from 'orbit/operation';
 import OperationProcessor from './operation-processor';
-import { toIdentifier, parseIdentifier } from './../lib/identifiers';
-import { operationType } from './../lib/operations';
+import { toIdentifier, parseIdentifier } from './../../lib/identifiers';
+import { operationType } from './../../lib/operations';
 
 /**
  An operation processor that ensures that a cache's data is consistent and
