+++ lib/orbit-common/cache/operation-processors/schema-consistency-processor.js
@@ -1,13 +1,13 @@
 import { isArray, isObject, isNone } from 'orbit/lib/objects';
 import Operation from 'orbit/operation';
 import OperationProcessor from './operation-processor';
-import { toIdentifier, parseIdentifier } from './../lib/identifiers';
+import { toIdentifier, parseIdentifier } from './../../lib/identifiers';
 import {
   operationType,
   addToHasManyOperation,
   replaceHasOneOperation,
   removeFromHasManyOperation
-} from './../lib/operations';
+} from './../../lib/operations';
 
 /**
  An operation processor that ensures that a cache's data is consistent with
