+++ lib/orbit-common/jsonapi-source.js
@@ -10,9 +10,9 @@ import Source from './source';
 import Serializer from './serializer';
 import JSONAPISerializer from './jsonapi-serializer';
 import { OperationNotAllowed, RecordNotFoundException, RecordAlreadyExistsException } from './lib/exceptions';
-import CacheIntegrityProcessor from 'orbit-common/operation-processors/cache-integrity-processor';
-import DeletionTrackingProcessor from 'orbit-common/operation-processors/deletion-tracking-processor';
-import SchemaConsistencyProcessor from 'orbit-common/operation-processors/schema-consistency-processor';
+import CacheIntegrityProcessor from 'orbit-common/cache/operation-processors/cache-integrity-processor';
+import DeletionTrackingProcessor from 'orbit-common/cache/operation-processors/deletion-tracking-processor';
+import SchemaConsistencyProcessor from 'orbit-common/cache/operation-processors/schema-consistency-processor';
 import { toIdentifier, parseIdentifier } from 'orbit-common/lib/identifiers';
 
 /**
